src/requirements.txt
PyOpenGL
glfw
numpy
Pillow 
